hello-llm/pyproject.toml
[project]
name = "hello-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.100.0",
    "python-dotenv>=1.2.2",
]
